refactor edit fragment and edit viewmodel to edit package
