--- app/src/main/java/app/stacq/plan/EditFragment.kt
@@ -1,31 +1,0 @@
-package app.stacq.plan
-
-import androidx.lifecycle.ViewModelProvider
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-class EditFragment : Fragment() {
-
-    companion object {
-        fun newInstance() = EditFragment()
-    }
-
-    private lateinit var viewModel: EditViewModel
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_edit, container, false)
-    }
-
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(EditViewModel::class.java)
-        // TODO: Use the ViewModel
-    }
-
-}
