+++ app/src/main/java/app/stacq/plan/ui/edit/EditViewModel.kt
@@ -1,4 +1,4 @@
-package app.stacq.plan
+package app.stacq.plan.ui.edit
 
 import androidx.lifecycle.ViewModel
