+++ app/src/main/res/layout/fragment_edit.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditFragment">
+    tools:context=".ui.edit.EditFragment">
 
     <TextView
         android:layout_width="match_parent"
